+++ package.json
@@ -104,8 +104,8 @@
   },
   "dependencies": {
     "@popperjs/core": "^2.4.4",
-    "@storybook/addon-cssresources": "^6.1.0-alpha.25",
-    "@storybook/addon-toolbars":  "^6.1.0-alpha.25",
+    "@storybook/addon-cssresources": "^6.1.0-alpha.30",
+    "@storybook/addon-toolbars": "^6.1.0-alpha.30",
     "invariant": "^2.2.4",
     "react-hook-form": "^6.8.2",
     "react-popper": "^2.2.3",
@@ -127,16 +127,16 @@
     "@semantic-release/exec": "^5.0.0",
     "@semantic-release/npm": "^7.0.6",
     "@semantic-release/release-notes-generator": "^9.0.1",
-    "@storybook/addon-a11y": "6.1.0-alpha.25",
-    "@storybook/addon-actions": "6.1.0-alpha.25",
-    "@storybook/addon-backgrounds": "^6.1.0-alpha.25",
-    "@storybook/addon-controls": "6.1.0-alpha.25",
-    "@storybook/addon-docs": "6.1.0-alpha.25",
-    "@storybook/addon-jest": "6.1.0-alpha.25",
-    "@storybook/addon-links": "6.1.0-alpha.25",
-    "@storybook/addons": "6.1.0-alpha.25",
-    "@storybook/react": "6.1.0-alpha.25",
-    "@storybook/theming": "6.1.0-alpha.25",
+    "@storybook/addon-a11y": "6.1.0-alpha.30",
+    "@storybook/addon-actions": "6.1.0-alpha.30",
+    "@storybook/addon-backgrounds": "^6.1.0-alpha.30",
+    "@storybook/addon-controls": "6.1.0-alpha.30",
+    "@storybook/addon-docs": "6.1.0-alpha.30",
+    "@storybook/addon-jest": "6.1.0-alpha.30",
+    "@storybook/addon-links": "6.1.0-alpha.30",
+    "@storybook/addons": "6.1.0-alpha.30",
+    "@storybook/react": "6.1.0-alpha.30",
+    "@storybook/theming": "6.1.0-alpha.30",
     "@tippy.js/react": "^3.1.1",
     "@typescript-eslint/parser": "^3.8.0",
     "@wfp/icons": "^0.0.82",
@@ -190,12 +190,11 @@
     "promise": "^8.1.0",
     "prop-types": "^15.7.2",
     "raw-loader": "^4.0.1",
-    "react": "^16.13.1",
+    "react": "^16.14.0",
     "react-dates": "^21.8.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "^16.14.0",
     "react-dropzone": "^11.0.2",
     "react-final-form": "^6.5.1",
-    "react-hooks": "^1.0.1",
     "react-is": "^16.13.1",
     "react-markdown": "^4.3.1",
     "react-prism": "^4.3.2",
