+++ src/components/Button/_button.scss
@@ -74,7 +74,7 @@
       var(--interactive-01),
       transparent,
       var(--inverse-01),
-      $hover-primary,
+      var(--hover-primary),
       var(--ui-01)
     );
   }
