+++ src/components/SecondaryNavigation/_secondary-navigation.scss
@@ -9,7 +9,7 @@
 
 @include exports('secondary-navigation') {
   .#{$prefix}--secondary-navigation {
-    background: var(ui-01);
+    background: var(--ui-01);
     @include reset;
     display: flex;
     border-bottom: 2px solid var(ui-02);
